Create UDFs of numpy functions supported by numba.
So far, do unary, binary, and monoid.  Semirings may come later.
This required modifying how UDFs are registered.  In particular, we are
now more relaxed about bool inputs that have non-bool outputs.

Numba has difficulty compiling functions for bools, so we upcast bool
values to int8 during the computation.  This may introduce subtle bugs
if coercions are not appropriate.  For example, bit-twidding and shifting
of bools may be incorrect!