Replace GbDelayed with typed expressions like MatrixExpression.
This PR touches lots of little things, and I move some things around such as the new `expr.py` file from `base.py`.
This is still a work in progress, and there are a couple rough spots to complete.
I plan to add nice repr to the `*Expression` objects.
I also plan to add more tests (gotta cover that code!).