Run `grblas.init` when testing installed version of grblas.
We need the top-level conftest.py for reasons, but this isn't currently available
when we test `grblas` when it's been installed.  So, let's move a couple pytest
setup functions to grblas/tests/conftest.py.  I don't know if this is considered
best practice, but it seems to check all the boxes for what I want.

When testing locally, we still have our nice options like --runslow and --nonblocking,
but when testing installed, we only test blocking suitesparse, and we don't skip slow.