Nicer repr for recorder and fix bug when in Notebook
Using a contextvar for the current recorder doesn't allow it to proceed across multiple
cells in Jupyter Notebook.  So, we fix this by keeping track of the previous recorder
and use that if the current recorder is None and hasn't been closed.

Nicer repr was done using `_repr_markdown_`, because syntax highlighting looks better.