Add `select` Operation (#207)
* select operation

Two ways of calling:
1. A.select(op, thunk)
2. gb.select.op(A, thunk)

When calling `A.select(op)`, the op may be passed as a string.

An additional set of "special" select functions take a comparator
expression. An example: `gb.select.row(A <= 2)`. Row, column, and value
are included, indicating what part of A is being compared in the select
statement.

* Fix import sorting

* Fixes based on feedback

- Use correct dtype based on thunk
- Handle Scalar thunk properly
- Rename IndexUnaryOp to SelectOp

* Fix for Windows

* Better handling of op.is_positional

* Updates based on feedback