Connection to Awkward Arrays (#280)
* Add io.to_awkward method

Converts a Vector or Matrix to an Awkward Array.
While an Awkward Array elegantly mimics pointers and values, it does
not naturally include the indices of a CSR matrix. However, Awkward
Arrays handle JSON-like structures of arrays, so we can easily add
values, indices, and even labels (in the case of hypersparse),
allowing a proper round trip journey from GraphBLAS to Awkward Arrays.

Awkward does not support iso-valued objects, so that will be expanded,
but yield equivalent information.

Awkward Array also does not currently support u64 for indexes, so we
are using i64. This should be fine as SuiteSparse:GraphBLAS only allows
indices up to 2**60, which means they fit in int64 even though GraphBLAS
internally stores them as uint64.